// frontend/gear-vault-web/src/api/generated/client/models/UpdateProfileDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateProfileDto = {
    name?: string | null;
    emailAddress?: string | null;
    username?: string | null;
    phoneNumber?: string | null;
    address?: string | null;
};